+++ Goblin/Models/Remind.cs
@@ -8,7 +8,7 @@ namespace Goblin.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int ID { get; set; }
 
-        public int VkID { get; set; }
+        public long VkID { get; set; }
         public DateTime Date { get; set; }
         public string Text { get; set; }
     }
